Expose commit and rollback to the transaction callback
This is the last step in supporting both explicit transaction commits
and rollbacks as well as the current implicit commits and rollbacks. If
a consuming application manually commits or rolls back a transaction,
`start` doesn't need to call it. If the application doesn't explicitly
commit or rollback a transaction, though, `start` will commit it or roll
it back. This maintains the current public API wile exposing more
functionality.